@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.11'
+gem 'rails', '4.0.2'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -19,13 +19,13 @@ gem 'devise'
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails'
+  gem 'coffee-rails'
 
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', :platforms => :ruby
 
-  gem 'uglifier', '>= 1.0.3'
+  gem 'uglifier'
 end
 
 gem 'jquery-rails'
@@ -45,7 +45,8 @@ gem 'jquery-rails'
 # To use debugger
 # gem 'debugger'
 gem "haml-rails"
-gem "formtastic", "~> 2.2.1"
-gem "awesome_nested_fields", "~> 0.6.1"
-gem 'bootstrap-sass', '~> 2.2.2.0'
-gem 'rails_admin'
+gem "formtastic"
+gem "awesome_nested_fields"
+gem 'bootstrap-sass'
+gem 'rails_admin', '~> 0.6.0', :git=>'git://github.com/sferik/rails_admin.git'
+gem 'protected_attributes'
